backend/requirements.txt
fastapi
uvicorn
PyMySQL
python-dotenv
pycryptodome
cryptography